# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HACKDAY'S API",
        "description": "Documentation API",
        "version": "1.0.0"
    },
    "paths": {
        "/admin/connect-user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Récupérer l'utilisateur connecté",
                "operationId": "12b149b7804b5ccc79aefba85a3f2789",
                "responses": {
                    "200": {
                        "description": "Utilisateur connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}